Move `Affine` and `Projective` to their own files
