+++ ec/src/models/short_weierstrass/affine.rs
@@ -1,0 +1,384 @@
+use ark_serialize::{
+    CanonicalDeserialize, CanonicalDeserializeWithFlags, CanonicalSerialize,
+    CanonicalSerializeWithFlags, SWFlags, SerializationError,
+};
+use ark_std::{
+    borrow::Borrow,
+    fmt::{Display, Formatter, Result as FmtResult},
+    io::{Read, Write},
+    ops::{Add, Mul, Neg, Sub},
+    rand::{
+        distributions::{Distribution, Standard},
+        Rng,
+    },
+    vec::Vec,
+    One, Zero,
+};
+
+use ark_ff::{fields::Field, PrimeField, ToConstraintField, UniformRand};
+
+use zeroize::Zeroize;
+
+use super::{Projective, SWCurveConfig};
+use crate::AffineRepr;
+
+/// Affine coordinates for a point on an elliptic curve in short Weierstrass
+/// form, over the base field `P::BaseField`.
+#[derive(Derivative)]
+#[derivative(
+    Copy(bound = "P: SWCurveConfig"),
+    Clone(bound = "P: SWCurveConfig"),
+    PartialEq(bound = "P: SWCurveConfig"),
+    Eq(bound = "P: SWCurveConfig"),
+    Debug(bound = "P: SWCurveConfig"),
+    Hash(bound = "P: SWCurveConfig")
+)]
+#[must_use]
+pub struct Affine<P: SWCurveConfig> {
+    #[doc(hidden)]
+    pub x: P::BaseField,
+    #[doc(hidden)]
+    pub y: P::BaseField,
+    #[doc(hidden)]
+    pub infinity: bool,
+}
+
+impl<P: SWCurveConfig> PartialEq<Projective<P>> for Affine<P> {
+    fn eq(&self, other: &Projective<P>) -> bool {
+        self.into_group() == *other
+    }
+}
+
+impl<P: SWCurveConfig> Display for Affine<P> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
+        match self.infinity {
+            true => write!(f, "infinity"),
+            false => write!(f, "({}, {})", self.x, self.y),
+        }
+    }
+}
+
+impl<P: SWCurveConfig> Affine<P> {
+    /// Constructs a group element from x and y coordinates.
+    /// Performs checks to ensure that the point is on the curve and is in the right subgroup.
+    pub fn new(x: P::BaseField, y: P::BaseField) -> Self {
+        let point = Self {
+            x,
+            y,
+            infinity: false,
+        };
+        assert!(point.is_on_curve());
+        assert!(point.is_in_correct_subgroup_assuming_on_curve());
+        point
+    }
+
+    /// Constructs a group element from x and y coordinates.
+    ///
+    /// # Warning
+    ///
+    /// Does *not* perform any checks to ensure the point is in the curve or
+    /// is in the right subgroup.
+    pub const fn new_unchecked(x: P::BaseField, y: P::BaseField) -> Self {
+        Self {
+            x,
+            y,
+            infinity: false,
+        }
+    }
+
+    pub const fn identity() -> Self {
+        Self {
+            x: P::BaseField::ZERO,
+            y: P::BaseField::ZERO,
+            infinity: true,
+        }
+    }
+
+    /// Attempts to construct an affine point given an x-coordinate. The
+    /// point is not guaranteed to be in the prime order subgroup.
+    ///
+    /// If and only if `greatest` is set will the lexicographically
+    /// largest y-coordinate be selected.
+    #[allow(dead_code)]
+    pub fn get_point_from_x(x: P::BaseField, greatest: bool) -> Option<Self> {
+        // Compute x^3 + ax + b
+        // Rust does not optimise away addition with zero
+        let x3b = if P::COEFF_A.is_zero() {
+            P::add_b(&(x.square() * &x))
+        } else {
+            P::add_b(&((x.square() * &x) + &P::mul_by_a(&x)))
+        };
+
+        x3b.sqrt().map(|y| {
+            let negy = -y;
+
+            let y = if (y < negy) ^ greatest { y } else { negy };
+            Self::new_unchecked(x, y)
+        })
+    }
+
+    /// Checks if `self` is a valid point on the curve.
+    pub fn is_on_curve(&self) -> bool {
+        if !self.infinity {
+            // Rust does not optimise away addition with zero
+            let mut x3b = P::add_b(&(self.x.square() * self.x));
+            if !P::COEFF_A.is_zero() {
+                x3b += &P::mul_by_a(&self.x);
+            };
+            self.y.square() == x3b
+        } else {
+            true
+        }
+    }
+}
+
+impl<P: SWCurveConfig> Affine<P> {
+    /// Checks if `self` is in the subgroup having order that equaling that of
+    /// `P::ScalarField`.
+    // DISCUSS Maybe these function names are too verbose?
+    pub fn is_in_correct_subgroup_assuming_on_curve(&self) -> bool {
+        P::is_in_correct_subgroup_assuming_on_curve(self)
+    }
+}
+
+impl<P: SWCurveConfig> Zeroize for Affine<P> {
+    // The phantom data does not contain element-specific data
+    // and thus does not need to be zeroized.
+    fn zeroize(&mut self) {
+        self.x.zeroize();
+        self.y.zeroize();
+        self.infinity.zeroize();
+    }
+}
+
+impl<P: SWCurveConfig> Distribution<Affine<P>> for Standard {
+    #[inline]
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Affine<P> {
+        loop {
+            let x = P::BaseField::rand(rng);
+            let greatest = rng.gen();
+
+            if let Some(p) = Affine::get_point_from_x(x, greatest) {
+                return p.mul_by_cofactor();
+            }
+        }
+    }
+}
+
+impl<P: SWCurveConfig> AffineRepr for Affine<P> {
+    type Config = P;
+    type BaseField = P::BaseField;
+    type ScalarField = P::ScalarField;
+    type Group = Projective<P>;
+
+    fn xy(&self) -> Option<(&Self::BaseField, &Self::BaseField)> {
+        (!self.infinity).then(|| (&self.x, &self.y))
+    }
+
+    #[inline]
+    fn generator() -> Self {
+        P::GENERATOR
+    }
+
+    fn from_random_bytes(bytes: &[u8]) -> Option<Self> {
+        P::BaseField::from_random_bytes_with_flags::<SWFlags>(bytes).and_then(|(x, flags)| {
+            // if x is valid and is zero and only the infinity flag is set, then parse this
+            // point as infinity. For all other choices, get the original point.
+            if x.is_zero() && flags.is_infinity() {
+                Some(Self::identity())
+            } else if let Some(y_is_positive) = flags.is_positive() {
+                Self::get_point_from_x(x, y_is_positive)
+                // Unwrap is safe because it's not zero.
+            } else {
+                None
+            }
+        })
+    }
+
+    fn mul_bigint(&self, by: impl AsRef<[u64]>) -> Self::Group {
+        P::mul_affine(self, by.as_ref())
+    }
+
+    /// Multiplies this element by the cofactor and output the
+    /// resulting projective element.
+    #[must_use]
+    fn mul_by_cofactor_to_group(&self) -> Self::Group {
+        P::mul_affine(self, Self::Config::COFACTOR)
+    }
+
+    /// Performs cofactor clearing.
+    /// The default method is simply to multiply by the cofactor.
+    /// Some curves can implement a more efficient algorithm.
+    fn clear_cofactor(&self) -> Self {
+        P::clear_cofactor(self)
+    }
+}
+
+impl<P: SWCurveConfig> Neg for Affine<P> {
+    type Output = Self;
+
+    /// If `self.is_zero()`, returns `self` (`== Self::zero()`).
+    /// Else, returns `(x, -y)`, where `self = (x, y)`.
+    #[inline]
+    fn neg(mut self) -> Self {
+        self.y = -self.y;
+        self
+    }
+}
+
+impl<P: SWCurveConfig, T: Borrow<Self>> Add<T> for Affine<P> {
+    type Output = Projective<P>;
+    fn add(self, other: T) -> Projective<P> {
+        let mut copy = self.into_group();
+        copy += other.borrow();
+        copy
+    }
+}
+
+impl<P: SWCurveConfig, T: Borrow<Self>> Sub<T> for Affine<P> {
+    type Output = Projective<P>;
+    fn sub(self, other: T) -> Projective<P> {
+        let mut copy = self.into_group();
+        copy -= other.borrow();
+        copy
+    }
+}
+
+impl<P: SWCurveConfig> Default for Affine<P> {
+    #[inline]
+    fn default() -> Self {
+        Self::identity()
+    }
+}
+
+impl<'a, P: SWCurveConfig, T: Borrow<P::ScalarField>> Mul<T> for Affine<P> {
+    type Output = Projective<P>;
+
+    #[inline]
+    fn mul(self, other: T) -> Self::Output {
+        self.mul_bigint(other.borrow().into_bigint())
+    }
+}
+
+// The projective point X, Y, Z is represented in the affine
+// coordinates as X/Z^2, Y/Z^3.
+impl<P: SWCurveConfig> From<Projective<P>> for Affine<P> {
+    #[inline]
+    fn from(p: Projective<P>) -> Affine<P> {
+        if p.is_zero() {
+            Affine::identity()
+        } else if p.z.is_one() {
+            // If Z is one, the point is already normalized.
+            Affine::new_unchecked(p.x, p.y)
+        } else {
+            // Z is nonzero, so it must have an inverse in a field.
+            let zinv = p.z.inverse().unwrap();
+            let zinv_squared = zinv.square();
+
+            // X/Z^2
+            let x = p.x * &zinv_squared;
+
+            // Y/Z^3
+            let y = p.y * &(zinv_squared * &zinv);
+
+            Affine::new_unchecked(x, y)
+        }
+    }
+}
+
+impl<P: SWCurveConfig> CanonicalSerialize for Affine<P> {
+    #[allow(unused_qualifications)]
+    #[inline]
+    fn serialize<W: Write>(&self, writer: W) -> Result<(), SerializationError> {
+        let (x, flags) = match self.infinity {
+            true => (P::BaseField::zero(), SWFlags::infinity()),
+            false => (self.x, SWFlags::from_y_sign(self.y > -self.y)),
+        };
+        x.serialize_with_flags(writer, flags)
+    }
+
+    #[inline]
+    fn serialized_size(&self) -> usize {
+        P::BaseField::zero().serialized_size_with_flags::<SWFlags>()
+    }
+
+    #[allow(unused_qualifications)]
+    #[inline]
+    fn serialize_uncompressed<W: Write>(&self, mut writer: W) -> Result<(), SerializationError> {
+        let (x, y, flags) = match self.infinity {
+            true => (
+                P::BaseField::zero(),
+                P::BaseField::zero(),
+                SWFlags::infinity(),
+            ),
+            false => (self.x, self.y, SWFlags::from_y_sign(self.y > -self.y)),
+        };
+        x.serialize(&mut writer)?;
+        y.serialize_with_flags(&mut writer, flags)?;
+        Ok(())
+    }
+
+    #[inline]
+    fn uncompressed_size(&self) -> usize {
+        // The size of the serialization is independent of the values
+        // of `x` and `y`, and depends only on the size of the modulus.
+        P::BaseField::zero().serialized_size()
+            + P::BaseField::zero().serialized_size_with_flags::<SWFlags>()
+    }
+}
+
+impl<P: SWCurveConfig> CanonicalDeserialize for Affine<P> {
+    #[allow(unused_qualifications)]
+    fn deserialize<R: Read>(reader: R) -> Result<Self, SerializationError> {
+        let (x, flags): (P::BaseField, SWFlags) =
+            CanonicalDeserializeWithFlags::deserialize_with_flags(reader)?;
+        if flags.is_infinity() {
+            Ok(Self::identity())
+        } else {
+            let p = Affine::<P>::get_point_from_x(x, flags.is_positive().unwrap())
+                .ok_or(SerializationError::InvalidData)?;
+            if !p.is_in_correct_subgroup_assuming_on_curve() {
+                return Err(SerializationError::InvalidData);
+            }
+            Ok(p)
+        }
+    }
+
+    #[allow(unused_qualifications)]
+    fn deserialize_uncompressed<R: Read>(
+        reader: R,
+    ) -> Result<Self, ark_serialize::SerializationError> {
+        let p = Self::deserialize_unchecked(reader)?;
+
+        if !p.is_in_correct_subgroup_assuming_on_curve() {
+            return Err(SerializationError::InvalidData);
+        }
+        Ok(p)
+    }
+
+    #[allow(unused_qualifications)]
+    fn deserialize_unchecked<R: Read>(mut reader: R) -> Result<Self, SerializationError> {
+        let x: P::BaseField = CanonicalDeserialize::deserialize(&mut reader)?;
+        let (y, flags): (P::BaseField, SWFlags) =
+            CanonicalDeserializeWithFlags::deserialize_with_flags(&mut reader)?;
+        match flags.is_infinity() {
+            true => Ok(Self::identity()),
+            false => Ok(Self::new_unchecked(x, y)),
+        }
+    }
+}
+
+impl<M: SWCurveConfig, ConstraintF: Field> ToConstraintField<ConstraintF> for Affine<M>
+where
+    M::BaseField: ToConstraintField<ConstraintF>,
+{
+    #[inline]
+    fn to_field_elements(&self) -> Option<Vec<ConstraintF>> {
+        let mut x = self.x.to_field_elements()?;
+        let y = self.y.to_field_elements()?;
+        let infinity = self.infinity.to_field_elements()?;
+        x.extend_from_slice(&y);
+        x.extend_from_slice(&infinity);
+        Some(x)
+    }
+}
